--- a/Changes
+++ b/Changes
@@ -267,6 +267,9 @@ _______________
   Falls back to using TransientTypeSet when marks are exhausted.
   (Jacques Garrigue and Takafumi Saikawa, review by Basile Clément)
 
+- #12946: Make generalization automatic when leaving scope
+  (Jacques Garrigue and Takafumi Saikawa, review by ???)
+
 - #12968: Attach location to constants in the parsetree
   (Jules Aguillon, review by Gabriel Scherer)