--- a/Changes
+++ b/Changes
@@ -270,7 +270,7 @@ _______________
 - #12946: Make generalization automatic when leaving scope.
   As a result, the `Ctype.generalize*` and `Ctype.correct_levels` functions
   were removed. The latter is now called `Ctype.duplicate_type`.
-  (Jacques Garrigue and Takafumi Saikawa, review by ???)
+  (Jacques Garrigue and Takafumi Saikawa, review by Richard Eisenberg)
 
 - #12968: Attach location to constants in the parsetree
   (Jules Aguillon, review by Gabriel Scherer)