--- a/toplevel/genprintval.ml
+++ b/toplevel/genprintval.ml
@@ -629,6 +629,6 @@ module Make(O : OBJ)(EVP : EVALPATH with type valu = O.t) = struct
             Oval_printer printer)
 
 
-    in nest tree_of_val max_depth obj (Ctype.correct_levels ty)
+    in nest tree_of_val max_depth obj ty
 
 end