--- a/typing/btype.mli
+++ b/typing/btype.mli
@@ -65,14 +65,6 @@ val lowest_level: int
 val with_new_pool: level:int -> (unit -> 'a) -> 'a * transient_expr list
         (* [with_new_pool ~level f] executes [f] and returns the nodes
            that were created at level [level] and above *)
-val register_last_pool: level:int -> unit
-        (* Register [level] so that its nodes are added to the same pool as
-           the previous level. Active until we leave an enclosing
-           [with_new_pool] or [with_last_pool]. *)
-val with_last_pool: level:int -> (unit -> 'a) -> 'a
-        (* [with_new_pool ~level f] executes [f] adding nodes at level [level]
-           to the pool of the previous level, so that they are returned by
-           the enclosing `with_new_pool` *)
 val add_to_pool: level:int -> transient_expr -> unit
         (* Add a type node to the pool associated to the level (which should
            be the level of the type node).