This PR makes the generalization automatic when leaving a scope.

Namely, we introduce a notion of type node pool in `Btype`.
* Every new type node is added to the pool corresponding to its level using `Btype.add_to_pool`.
* `Ctype.with_local_level_generalize[_structure]` create a new pool.
**  When leaving its scope, all the nodes in the pool which have level higher than then previous level are generalized
** In the `_structure` case, type variables rather have their level lowered.
** Nodes with a level lower or equal to the previous level are moved to the pool corresponding to their actual level.
** Nodes that are already at `generic_level` and `Tlink`'s are ignored.
** Calls to `lower_contravariant` must be added to the `before_generalize` argument.
* For compatibility, if there is no pool for a given level, then the nodes are added to the last pool.
* Since nodes at `generic_level` are not tracked, we need to add them back to a pool  if their level is lowered (in `update_level`).

We could get rid of all calls to `generalize` and `generalize_structure`, but `limited_generalize` and `generalize_spine` are still needed for classes.
Checked that there is no negative impact on compiler performance on `ocamldoc`.

Written with @t6s