@gasche 2024-02-13
Naive question: to perform generalization, we only *need* to traverse the type variables that occur in the result type. Here, if I understand correctly, we generalize all type variables at the level we are exiting, including those that were only used internally for type-checking and do not occur in the result. This changes the algorithmic complexity of inference, right?

In the past I proposed that the `with_local_level_*` wrappers could pass a function argument that "registers" a type expression as being part of the result, which marks it for generalization later. This would be less cumbersome than the current code in trunk, still more work than the present proposal in the PR, higher risk of introducing bugs by forgetting to register certain variables, but it would have the right algorithmic complexity. Maybe it would be possible to have a debug mode that tracks all type variables in a pool, but invalidates all those that have not been "recorded" (registered as roots) in this way, and would help catch this family of bugs (forgetting to register some types).

For example, in trunk we have

```ocaml
      let cty, ty, force, cty', ty', force' =
        with_local_level_iter ~post:generalize_structure begin fun () ->
          let (cty, ty, force) =
            Typetexp.transl_simple_type_delayed env sty
          and (cty', ty', force') =
            Typetexp.transl_simple_type_delayed env sty'
          in
          ((cty, ty, force, cty', ty', force'),
           [ ty; ty' ])
        end
```

with your PR we have the nicer

```ocaml
      let cty, ty, force, cty', ty', force' =
        with_local_level_generalize_structure begin fun () ->
          let (cty, ty, force) =
            Typetexp.transl_simple_type_delayed env sty
          and (cty', ty', force') =
            Typetexp.transl_simple_type_delayed env sty'
          in
          (cty, ty, force, cty', ty', force')
        end
```

but a middle-road would be

```ocaml
      let cty, ty, force, cty', ty', force' =
        with_local_level_iter begin fun gen ->
          let (cty, ty, force) =
            Typetexp.transl_simple_type_delayed env sty
          and (cty', ty', force') =
            Typetexp.transl_simple_type_delayed env sty'
          in
          (cty, gen.structure ty, force,
           cty', gen.structure ty', force')
        end
```