@garrigue 2024-02-15
Indeed, it looks like we do useless work when we generalize nodes that will just be garbage collected.
However, our measurements show that there is no impact on performance, so this is not a big deal.
Moreover, the meaning of "being part of the result" is not so clear.
In trunk, we only generalize nodes which are still relevant to type inference. But this means that we break level-decreasing invariants on the types inside the typed tree, so that we have to take fresh copies when we use them in the middle-end and back-end (hence the comment about using `Ctype.correct_levels`). With automatic generalize, it should be possible to dispense with that, as we always generalize all the nodes that can be. (Being sure that the invariant is always enforced may require a more thorough review.)