@garrigue 2024-02-20
Performance measurements on ocamldoc, using ocamlc.opt and ocamlopt.opt.
There seems to be a 2% slowdown with ocamlc.opt, but with ocamlopt.opt it is not even clear which is faster.
We also did some measurements on versions that generalize a bit less for reference (correspond to log messages in git).

```
ocamlc.opt
trunk
1.979u 0.913s 0:03.02 95.3%	0+0k 0+0io 18159pf+0w
1.989u 0.903s 0:03.00 96.0%	0+0k 0+0io 17930pf+0w
1.984u 0.905s 0:03.01 95.6%	0+0k 0+0io 17930pf+0w

automatic_generalize
2.011u 0.938s 0:03.06 96.0%	0+0k 0+0io 17691pf+0w
2.021u 0.921s 0:03.05 96.3%	0+0k 0+0io 17691pf+0w
2.023u 0.932s 0:03.07 96.0%	0+0k 0+0io 17691pf+0w

register pools for all levels
2.011u 0.925s 0:03.05 96.0%	0+0k 0+0io 17839pf+0w
2.010u 0.932s 0:03.05 96.3%	0+0k 0+0io 17839pf+0w
2.013u 0.933s 0:03.06 96.0%	0+0k 0+0io 17839pf+0w

automatic generalization in typedecl
2.002u 0.917s 0:03.03 96.0%	0+0k 0+0io 17928pf+0w
1.999u 0.911s 0:03.03 95.7%	0+0k 0+0io 17928pf+0w
2.000u 0.911s 0:03.02 96.3%	0+0k 0+0io 17928pf+0w

ocamlopt.opt
trunk
4.510u 1.438s 0:06.52 91.1%	0+0k 0+0io 30239pf+0w
4.507u 1.414s 0:06.46 91.4%	0+0k 0+0io 30007pf+0w
4.510u 1.405s 0:06.46 91.4%	0+0k 0+0io 30007pf+0w
4.506u 1.413s 0:06.45 91.6%	0+0k 0+0io 30007pf+0w

automatic_generalize
4.495u 1.381s 0:06.39 91.8%	0+0k 0+0io 29747pf+0w
4.512u 1.415s 0:06.47 91.4%	0+0k 0+0io 29747pf+0w
4.518u 1.412s 0:06.47 91.4%	0+0k 0+0io 29747pf+0w
4.491u 1.413s 0:06.48 91.0%	0+0k 0+0io 29747pf+0w
```