@garrigue 2024-02-20
The PR is now ready for review.
As the measurements show ,there seems to be a small impact of performance, but we believe the simplicity of use and the improved invariant are worth it. There seems to be no complexity problem, as nodes that do not escape a scope are going to be only generalized once, and at low cost.