@garrigue 2024-03-05
I did try using a hash table (branch `automatic_generalize_hashtbl`).

The results are not bad:
```
9.612u 3.409s 0:13.26 98.1%	0+0k 0+0io 96191pf+0w
9.688u 3.525s 0:13.44 98.2%	0+0k 0+0io 96191pf+0w
9.644u 3.429s 0:13.27 98.4%	0+0k 0+0io 96191pf+0w
```
I.e., this is very close to trunk. Could switch to this if the difference matters.