@garrigue 2024-04-18
Removed `Ctype.correct_levels`. This didn't take long, and the test suite goes through. You might still want to check on a bigger code base.