@garrigue 2024-04-18
I also checked that remove `Ctype_correct_levels` from `Typeop.scrape_ty` indeed fails in trunk (otherwise, the test suite is dubious...)
There were surprisingly few failures, but the following tests indeed fail (in trunk):
```
List of failed tests:
    tests/effects/issue479.ml
    tests/typing-misc/constraints.ml
    tests/typing-objects/Exemples.ml
    tests/typing-poly/poly.ml
```
Those are the ones exercising the type checker.
There were no errors in `typing-gadts`.
By the way, `automatic_generalize` fails on `tests/lib-threads/torture.ml`, but IIRC this one is stochastic.