@garrigue 2024-05-30
@goldfirere Thank for your view review. I am still updating the code, as you helped discover a number of possible improvements, and hope to finish by next week. Then we can schedule a call.