@garrigue 2024-06-19
Thank you for your approval. However, rebasing causes an error in the testsuite, and we are still inverstigating the cause. (There is an easy workaround, but I would really like to understand what is happening.)