@garrigue 2024-06-20
@goldfirere Ironically, it seems that cause is #13088, your PR about making `list_labels` pure, that I reviewed myself.
What is happening is not completely clear, but this shows that it is very difficult to produce a correct map from levels to pools if we work level by level.
So we tried another approach, using rather a stack of pools, associated the level at which they will be generalized. Then one just add nodes to the topmost pool whose level is sufficiently low. Search in this structure is linear, but since insertion is amortized, this ends up being more efficient than lookup in a map. And the correctness of lookup in this structure is trivial.

It would be nice if you could review the last commit, which is the only new one. It also uses `with_level` in a few places, to avoid direct assignments to `current_level`.

We again did benchmark compiling `ocamldoc` and `ocamlopt` using `ocamlc.opt`, and divine surprise, we are now faster than trunk for `ocamlopt`.
```text
ocamlopt
trunk
time make ocamlopt CAMLC=../trunk/ocamlc.opt > /dev/null
9.924u 3.684s 0:13.86 98.1%	0+0k 0+0io 99022pf+0w
10.001u 3.868s 0:14.19 97.6%	0+0k 0+0io 98814pf+0w
9.943u 3.717s 0:13.87 98.4%	0+0k 0+0io 98582pf+0w

with pool stack
time make ocamlopt CAMLC=./ocamlc.agps.opt > /dev/null
9.856u 3.728s 0:13.80 98.3%	0+0k 0+0io 98786pf+0w
9.882u 3.688s 0:13.77 98.4%	0+0k 0+0io 98553pf+0w
9.946u 3.908s 0:14.16 97.7%	0+0k 0+0io 98553pf+0w

ocamldoc
trunk
1.732u 0.715s 0:02.50 97.6%	0+0k 0+0io 17914pf+0w
1.739u 0.717s 0:02.51 97.2%	0+0k 0+0io 17913pf+0w
1.742u 0.716s 0:02.51 97.6%	0+0k 0+0io 17913pf+0w

with pool stack (ocamlc.agps.opt)
1.738u 0.719s 0:02.51 97.2%	0+0k 0+0io 17913pf+0w
1.744u 0.720s 0:02.52 97.6%	0+0k 0+0io 17913pf+0w
1.753u 0.724s 0:02.55 96.8%	0+0k 0+0io 17913pf+0w
```